+++ framework/domain/src/main/java/is/codion/framework/domain/entity/attribute/DefaultForeignKeyDefinition.java
@@ -41,7 +41,6 @@ final class DefaultForeignKeyDefinition extends AbstractAttributeDefinition<Enti
 
 	private final Set<Column<?>> readOnlyColumns;
 	private final List<Attribute<?>> attributes;
-	private final boolean nullable;
 	private final int referenceDepth;
 	private final boolean soft;
 
@@ -49,7 +48,6 @@ final class DefaultForeignKeyDefinition extends AbstractAttributeDefinition<Enti
 		super(builder);
 		this.readOnlyColumns = builder.readOnlyColumns;
 		this.attributes = builder.attributes;
-		this.nullable = builder.nullable;
 		this.referenceDepth = builder.referenceDepth;
 		this.soft = builder.soft;
 	}
@@ -57,11 +55,6 @@ final class DefaultForeignKeyDefinition extends AbstractAttributeDefinition<Enti
 	@Override
 	public ForeignKey attribute() {
 		return (ForeignKey) super.attribute();
-	}
-
-	@Override
-	public boolean nullable() {
-		return nullable;
 	}
 
 	@Override
@@ -108,19 +101,12 @@ final class DefaultForeignKeyDefinition extends AbstractAttributeDefinition<Enti
 		private final EntityType referencedType;
 
 		private List<Attribute<?>> attributes = emptyList();
-		private boolean nullable;
 		private boolean soft = false;
 		private int referenceDepth = REFERENCE_DEPTH.getOrThrow();
 
 		DefaultForeignKeyDefinitionBuilder(ForeignKey foreignKey) {
 			super(foreignKey);
 			this.referencedType = foreignKey.referencedType();
-		}
-
-		@Override
-		public ForeignKeyDefinition.Builder nullable(boolean nullable) {
-			this.nullable = nullable;
-			return this;
 		}
 
 		@Override